changement type variable note pour accepter des réels

+++ exercice 4.c
@@ -1,13 +1,13 @@
 #include <stdio.h>;
 
 int main(){
-    int n = 1;
+    float n = 1;
     int nbr_note = 0;
     float somme = 0;
     float moyenne;
     while(n != -1){
         printf("saisir note");
-        scanf("%d", &n);
+        scanf("%f", &n);
         if (n != -1){
             nbr_note++;
             somme = somme + n;
